REFACTOR : 상품 수정 페이지 타입 변경 및 쿼리 적용
