+++ src/hooks/queries/products/useHideMutation.ts
