+++ src/pages/productPage/main/index.tsx
@@ -14,40 +14,14 @@ import logo from 'assets/logo/logo.svg';
 import notifications from 'assets/icons/notifications.svg';
 import { useEffect } from 'react';
 import { useSearchActions } from 'store/search';
-import { instance } from 'apis';
-import { useActiveCategory, useClickedOnSale, useProductListActions } from 'store/productListData';
-import { useQuery } from '@tanstack/react-query';
-
-async function getProductListData(category: string, onSale: string | null) {
-  try {
-    const response = await instance.get(
-      `/products/category?categoryName=${category}&postStatus=${onSale}&pageNumber=0`
-    );
-    console.log('물품 리스트 불러오기 성공', response);
-    return response.data.content;
-  } catch (e) {
-    console.log('물품 리스트 불러오기 실패 ', e);
-  }
-}
+import { useProductListActions } from 'store/productListData';
+import { useProductMainQuery } from 'hooks/queries/products/useProductMainQuery';
 
 const ProductMain = () => {
   const navigate = useNavigate();
-  const clickedOnSale = useClickedOnSale();
+  const productMainQuery = useProductMainQuery();
   const { changeSearchData } = useSearchActions();
-  const activeCategory = useActiveCategory();
   const { setActiveCategory } = useProductListActions();
-  const { setInitialProductList } = useProductListActions();
-
-  const { data, error, isLoading } = useQuery({
-    queryKey: ['products', activeCategory, clickedOnSale],
-    queryFn: () => getProductListData(activeCategory, clickedOnSale),
-  });
-
-  useEffect(() => {
-    if (data) {
-      setInitialProductList(data);
-    }
-  }, [data]);
 
   useEffect(() => {
     changeSearchData('');
@@ -75,8 +49,8 @@ const ProductMain = () => {
         </section>
 
         <section className="items">
-          {error && <S.Error>상품을 불러오지 못했습니다.</S.Error>}
-          {isLoading ? <Loading /> : <ListTradeItems />}
+          {productMainQuery.error && <S.Error>상품을 불러오지 못했습니다.</S.Error>}
+          {productMainQuery.isLoading ? <Loading /> : <ListTradeItems />}
         </section>
 
         <Link to={'/product/new'}>
