+++ src/pages/productPage/product-edit/index.tsx
@@ -2,45 +2,18 @@ import { FormTrade, Header } from 'components/index';
 import * as S from './styled';
 import close from 'assets/icons/close_large.svg';
 import { useNavigate, useParams } from 'react-router-dom';
-import { useMutation, useQuery } from '@tanstack/react-query';
-import { instance } from 'apis';
-import { userId } from 'data/shared';
-import { TradeFormData } from 'types/types';
 import { useFormData, useFormDataActions } from 'store/formData';
 import { useEffect } from 'react';
-
-const getProductDetailData = async (id: string | undefined) => {
-  try {
-    const response = await instance.get(`/products/${id}`);
-    console.log('데이터 가져오기 성공', response);
-    return response.data;
-  } catch (e) {
-    console.log('데이터 가져오기 실패', e);
-  }
-};
+import { useProductDetailQuery } from 'hooks/queries/products/useProductDetailQuery';
+import { useProductEditMutation } from 'hooks/queries/products/useProductEditMutation';
 
 const ProductEdit = () => {
   const { id } = useParams();
   const navigate = useNavigate();
   const formData = useFormData();
   const { receiveData, changeImgFileToString } = useFormDataActions();
-  const productDetailQuery = useQuery({
-    queryKey: ['products', 'product-detail'],
-    queryFn: () => getProductDetailData(id),
-  });
-
-  const productEditMutation = useMutation({
-    mutationFn: (sendData: TradeFormData) =>
-      instance.put(`/products/edit/${userId}/${id}`, sendData),
-    onSuccess: (res) => {
-      console.log('상품 수정에 성공했습니다.', res);
-      alert('상품 수정에 성공했습니다.');
-    },
-    onError: (error) => {
-      console.log('상품 수정에 실패했습니다.', error);
-      alert('상품 수정에 실패했습니다.');
-    },
-  });
+  const productDetailQuery = useProductDetailQuery(id as string);
+  const { mutate: productEditMutation } = useProductEditMutation(id as string);
 
   useEffect(() => {
     receiveData(productDetailQuery.data);
@@ -63,7 +36,7 @@ const ProductEdit = () => {
       <S.Content>
         <FormTrade
           formData={formData}
-          handleSubmitAction={productEditMutation.mutate}
+          handleSubmitAction={() => productEditMutation(formData)}
           btnText={'수정 완료'}
         />
       </S.Content>
